pyproject.toml
[project]
name = "iSparrow"
version = "0.1"
requires-python = ">= 3.9"
authors=[{name = "Inga Ulusoy", email = "inga.ulusoy@uni-heidelberg.de"}, {name = "Harald Mack", email = "harald.mack@uni-heidelberg.de"}]
description = "Smart-IoT-based Passive bioAcoustic monitoRing of site-specific co-pResence of zOonotic Wildlife hosts and humans"
readme ="README.md"
license={file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache License",
]

dependencies = [
    "numpy", 
    "librosa", 
    "tensorflow", 
    "PyYAML", 
    "birdnetlib==0.15.0", 
    "pooch", 
    "resampy", # for audio loading
    "tensorflow-hub",
    "validators", # test url validity 
    "platformdirs", # get cache dirs without os dependence etc
    "ffmpeg-python"
]

[project.urls]
Repository = "https://github.com/ssciwr/iSparrow"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"



# add additional stuff when needed, poetry support? 